# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BarberMetrics API specification
servers:
  - url: /api
    description: Base API path
...
security:
  - bearerAuth: []
paths:
...
  /dashboard/monthly-analysis:
    get:
      operationId: getMonthlyAnalysis
      tags: [dashboard]
      summary: Get monthly performance analysis
      parameters:
        - name: month
          in: query
          description: "Month in YYYY-MM format, defaults to current month"
          schema:
            type: string
      responses:
        "200":
          description: Monthly performance analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyAnalysis"
...
  /settings/daily-goal:
    get:
      operationId: getDailyGoal
      tags: [dashboard]
      summary: Get daily earnings goal
      responses:
        "200":
          description: Daily goal setting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyGoal"
    put:
      operationId: updateDailyGoal
      tags: [dashboard]
      summary: Update daily earnings goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDailyGoalBody"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyGoal"
  /settings/work-hours:
    get:
      operationId: getWorkHours
      tags: [dashboard]
      summary: Get work hours settings
      responses:
        "200":
          description: Work hours
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkHours"
    put:
      operationId: updateWorkHours
      tags: [dashboard]
      summary: Update work hours settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkHours"
      responses:
        "200":
          description: Updated work hours
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkHours"
  /settings/commission:
    get:
      operationId: getCommission
      tags: [dashboard]
      summary: Get barber commission percentage
      responses:
        "200":
          description: Commission setting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommissionSetting"
    put:
      operationId: updateCommission
      tags: [dashboard]
      summary: Update barber commission percentage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommissionSetting"
      responses:
        "200":
          description: Updated commission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommissionSetting"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CommissionSetting:
      type: object
      properties:
        commissionPercent:
          type: number
          minimum: 0
          maximum: 100
      required:
        - commissionPercent
...
    DailyGoal:
      type: object
      properties:
        goal:
          type: number
      required:
        - goal
    UpdateDailyGoalBody:
      type: object
      properties:
        goal:
          type: number
      required:
        - goal
    WorkHours:
      type: object
      properties:
        hoursPerDay:
          type: number
        daysPerWeek:
          type: integer
      required:
        - hoursPerDay
        - daysPerWeek
    DayEarnings:
      type: object
      properties:
        date:
          type: string
        earnings:
          type: number
        appointmentCount:
          type: integer
      required:
        - date
        - earnings
        - appointmentCount
    HourStat:
      type: object
      properties:
        hour:
          type: integer
        count:
          type: integer
      required:
        - hour
        - count
    WeekdayStat:
      type: object
      properties:
        weekday:
          type: integer
        dayName:
          type: string
        avgEarnings:
          type: number
        count:
          type: integer
      required:
        - weekday
        - dayName
        - avgEarnings
        - count
    ServiceRankItem:
      type: object
      properties:
        service:
          type: string
        count:
          type: integer
        totalEarnings:
          type: number
      required:
        - service
        - count
        - totalEarnings
    MonthlyAnalysis:
      type: object
      properties:
        month:
          type: string
        workedDays:
          type: integer
        totalEarnings:
          type: number
        dailyAverage:
          type: number
        monthlyForecast:
          type: number
        topDays:
          type: array
          items:
            $ref: "#/components/schemas/DayEarnings"
        busiestHours:
          type: array
          items:
            $ref: "#/components/schemas/HourStat"
        weekdayStats:
          type: array
          items:
            $ref: "#/components/schemas/WeekdayStat"
        serviceRanking:
          type: array
          items:
            $ref: "#/components/schemas/ServiceRankItem"
      required:
        - month
        - workedDays
        - totalEarnings
        - dailyAverage
        - monthlyForecast
        - topDays
        - busiestHours
        - weekdayStats
        - serviceRanking